AIOserver/requirements.txt
asgiref
bson
coreapi-cli
dataclasses
dj-database-url
django-address
django-admin
django-allauth
django-better-admin-arrayfield
django-cors-headers
django-environ
django-extensions
django-filter
django-flags
django-friendship
django-jsonfield
django-resized
django-rest-auth
django-rest-framework
django-snippet
django-tags-input
django-waffle
djangorestframework-bulk
djangorestframework-jwt
drf-dynamic-fields
drf-nested-serializer
drf-writable-nested
grpcio
gunicorn
joblib
jsonfield
markdown
msedge-selenium-tools
pillow
PyJWT==1.7.1
psycopg2
psycopg2-binary
scipy
tabulate
virtualenv
webdriver-manager
whitenoise
